Move cache client to `common` module
