+++ common/client/cache/dummycache.go
@@ -2,8 +2,6 @@ package cache
 
 import (
 	"context"
-
-	errs "github.com/wormhole-foundation/wormhole-explorer/api/internal/errors"
 )
 
 // DummyCacheClient dummy cache client.
@@ -18,7 +16,7 @@ func NewDummyCacheClient() *DummyCacheClient {
 // Get get method is a dummy method that always does not find the cache.
 // Use this Get function when run development enviroment
 func (d *DummyCacheClient) Get(ctx context.Context, key string) (string, error) {
-	return "", errs.ErrNotFound
+	return "", ErrNotFound
 }
 
 // Close dummy cache client.
